move version matching to prepare
