+++ package.json
@@ -38,6 +38,7 @@
     "test:chromatic": "if-env TRAVIS_EVENT_TYPE=push && yarn chromatic || echo 'Skip chromatic'",
     "chromatic":
       "chromatic test --storybook-addon --script-name=dev:sandbox --exit-zero-on-changes",
+    "prepare": "node tooling/prepare.js",
     "deploy": "node tooling/publish.js",
     "codemods": "jscodeshift -t src/codemods src/"
   },
