+++ src/components/atoms/icon/icons.json
@@ -182,6 +182,13 @@
       "width": 20,
       "height": 20
     },
+    "play": {
+      "paths": [
+        "M6.11 16.837v.152h-.237L5 17.544V2l12.213 7.772L6.11 16.837zm0-1.143l8.882-5.922L6.11 3.85v11.844z"
+      ],
+      "width": 20,
+      "height": 20
+    },
     "plus": {
       "paths": [
         "M16.988 10.328h-6.661v6.662a.555.555 0 0 1-1.11 0v-6.662H2.554a.555.555 0 0 1 0-1.111h6.661V2.555a.555.555 0 0 1 1.11 0v6.662h6.662a.555.555 0 0 1 0 1.111z"
